[benc/item-flipbook] Rename ItemSelection to ItemDisplay
