+++ packages/perseus/src/testing/item-flipbook/item-flipbook.tsx
@@ -1,54 +1,14 @@
 import * as React from "react";
 import {useReducer, useRef} from "react";
 
-import {ServerItemRenderer} from "../../server-item-renderer";
-import {testDependenciesV2} from "../test-dependencies";
-
 import {createItemFlipbookModel} from "./item-flipbook-model";
 
-import type {PerseusItemSelection, ViewModel} from "./item-flipbook-view-model";
+import {ItemFlipbookView} from "./item-flipbook-view";
 
 export function ItemFlipbook() {
     const model = useModel(createItemFlipbookModel);
 
     return <ItemFlipbookView viewModel={model.present()} />;
-}
-
-interface ViewProps {
-    viewModel: ViewModel;
-}
-
-function ItemFlipbookView({viewModel}: ViewProps) {
-    return (
-        <>
-            <textarea
-                value={viewModel.textareaValue}
-                onChange={(e) => viewModel.setTextareaValue(e.target.value)}
-            />
-            <ItemSelectionView itemSelection={viewModel.itemSelection} />
-        </>
-    );
-}
-
-interface ItemSelectionViewProps {
-    itemSelection: PerseusItemSelection;
-}
-
-function ItemSelectionView({itemSelection}: ItemSelectionViewProps) {
-    switch (itemSelection.status) {
-        case "no-item":
-            return <p>No item selected for viewing.</p>;
-        case "parse-error":
-            // TODO: use Wonder Blocks
-            return <p>{itemSelection.message}</p>;
-        case "item":
-            return (
-                <ServerItemRenderer
-                    item={itemSelection.item}
-                    dependencies={testDependenciesV2}
-                />
-            );
-    }
 }
 
 function useModel<T>(constructor: (observer: () => void) => T): T {
